+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Camera auto-focus in preview mode
+- Support for nested scopes using `{}` syntax
 
 ## [v0.0.3]